requirements.txt
browser-use==0.1.29
crewai==0.98.0
langchain-community==0.3.15
langchain-openai==0.3.1
nest-asyncio==1.6.0
pydantic==2.10.6
python-dotenv==1.0.1
PyYAML==6.0.1

